+++ include/oys_swap_chain.hpp
@@ -23,9 +23,9 @@ namespace oys {
         OysSwapChain(const OysSwapChain&) = delete;
         OysSwapChain& operator=(const OysSwapChain&) = delete;
 
-        VkFramebuffer getFrameBuffer(int index) { return swapChainFramebuffers[index]; }
+        VkFramebuffer getFrameBuffer(size_t index) { return swapChainFramebuffers[index]; }
         VkRenderPass getRenderPass() { return renderPass; }
-        VkImageView getImageView(int index) { return swapChainImageViews[index]; }
+        VkImageView getImageView(size_t index) { return swapChainImageViews[index]; }
         size_t imageCount() { return swapChainImages.size(); }
         VkFormat getSwapChainImageFormat() { return swapChainImageFormat; }
         VkExtent2D getSwapChainExtent() { return swapChainExtent; }
